requirements.txt
django
djangorestframework
konlpy
nltk
pandas
scikit-learn
numpy
Pillow
